AI_Resume_Screening_System/requirements.txt
flask
spacy
gunicorn
python-docx
pdfplumber